pyproject.toml
[project]
name = "mibudge"
version = "0.1.0"
description = "Personal budgeting service inspired by Simple Bank"
requires-python = ">=3.13"
dependencies = [
    # Core
    "django>=5.2,<6",
    "djangorestframework",
    "django-environ",
    "django-extensions",
    "django-filter",
    "django-model-utils",
    "django-cors-headers",
    # Auth
    "djangorestframework-simplejwt",
    "django-allauth",
    "django-guardian",
    "argon2-cffi",
    # Money & scheduling
    "django-money",
    "django-recurrence",
    # Celery
    "celery",
    "django-celery-beat",
    "flower",
    # Cache & broker
    "redis[hiredis]",
    "django-redis",
    # Database
    "psycopg[binary]",
    # Static files & media
    "whitenoise",
    "django-vite",
    "Pillow",
    # Server
    "uvicorn[standard]",
    # Service readiness
    "wait-for-it",
    # Utilities
    "python-slugify",
    "python-dateutil",
    "markdown",
    "PyYAML",
    "Pygments",
    "rich",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app/moneypools", "app/users"]

[dependency-groups]
dev = [
    # Testing
    "pytest",
    "pytest-django",
    "pytest-factoryboy",
    "pytest-sugar",
    "pytest-cov",
    "factory-boy",
    "faker",
    "fakeredis",
    "pytest-mock",
    "freezegun",
    # Type checking
    "django-stubs[compatible-mypy]>=5.2,<6",
    # Linting & formatting
    "ruff",
    "pre-commit",
    # Debugging
    "ipdb",
    "ipython",
    "werkzeug",
    # Utilities
    "django-debug-toolbar",
]

########################################################################
# Tool configurations
########################################################################

[tool.ruff]
target-version = "py313"
line-length = 80
exclude = ['migrations/.+\.py$']

[tool.ruff.lint]
select = ["E", "W", "F", "B", "C4", "UP", "I"]
ignore = ["E501", "B008", "B905"]

[tool.ruff.lint.isort]
known-first-party = ["moneypools", "users", "config"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
exclude = ['__pycache__', 'migrations', '.venv', 'venv*']
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
# django-manager-missing is a known false-positive with djmoney (MoneyManager)
disable_error_code = ["django-manager-missing"]

[tool.django-stubs]
django_settings_module = "config.settings"
strict_settings = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
addopts = "--durations=10 --durations-min=0.1"
testpaths = ["app/tests"]
pythonpath = ["app"]
markers = [
    "smoke: marks tests as smoke",
    "integration",
]
